requirements.txt
langchain>=0.1.0
langchain-anthropic>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
trustcall>=0.1.0
